Move type_cache into its own module; refactor out kind caches
