+++ concept/type_/type_cache/mod.rs
@@ -1,0 +1,11 @@
+/*
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at https://mozilla.org/MPL/2.0/.
+ */
+
+mod type_cache;
+mod kind_cache;
+mod selection;
+
+pub use type_cache::TypeCache;
