+++ concept/type_/type_cache/selection.rs
@@ -5,7 +5,8 @@
  */
 
 use crate::type_::{
-    type_cache::{CommonTypeCache, OwnerPlayerCache, TypeCache},
+    type_cache::TypeCache,
+    type_cache::kind_cache::{CommonTypeCache, OwnerPlayerCache},
     type_manager::KindAPI,
 };
